Intelligent Crop Disease Classifier/requirements.txt
tensorflow
gradio
Pillow
numpy
